+++ hazelcast/src/main/java/com/hazelcast/internal/management/events/AbstractWanConfigurationEventBase.java
@@ -20,7 +20,7 @@ import com.hazelcast.internal.json.JsonObject;
 
 import java.util.UUID;
 
-public abstract class AbstractWanConfigurationEventBase extends AbstractWanEvent {
+public abstract class AbstractWanConfigurationEventBase extends AbstractIdentifiedEvent {
     private final String wanConfigName;
 
     protected AbstractWanConfigurationEventBase(UUID uuid, String wanConfigName) {
